requirements.txt
streamlit==1.29.0
pandas==2.1.3
Pillow==10.1.0
pytesseract==0.3.10
google-generativeai==0.3.1
python-dotenv==1.0.0
opencv-python==4.9.0.80
matplotlib==3.8.3
requests==2.31.0 